// sanity.types.ts
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top: number;
  bottom: number;
  left: number;
  right: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x: number;
  y: number;
  height: number;
  width: number;
};
// ...
export type Slug = {
  _type: "slug";
  current: string;
  source?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;

// Source: sanity/helpers/index.ts
// Variable: PRODUCTS_QUERY
// Query: *[_type=="product"] | order(name asc){    ...,"categories": coalesce(			categories[]->,			[])  }
export type PRODUCTS_QUERY_RESULT = Array<{
  _id: string;
  _type: "product";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name: string;
  slug: Slug;
  image?: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  description?: string;
  price: number;
  discount: number;
  categories:
    | Array<{
        _id: string;
        _type: "category";
        _createdAt: string;
        _updatedAt: string;
        _rev: string;
        title: string;
        slug: Slug;
        description?: string;
      }>
    | Array<never>;
  stock?: number;
  label?: string;
  status?: "hot" | "new" | "sale";
}>;